Initial commit
Connected to the AlphaAdvantage stock index API, parsed the JSON data into custom structs with custom Decoder keys. Created the basic logic for the main viewController to display limited data (ticker/volume/pricing) in a tableView cell. Storyboard file is only temporary at this point to ensure that the tableView is properly displaying the data.