Commit of detailed view controller
This is the initial commit with the DetailViewController, Storyboard Files and model file. No logic is built into the VC at this point, just the basic framework for how It's going to look. I also made changes to the MainViewController that forces devices with "regular" widths (iPad and iPhone+ in landscape) to display both the master and detail view controller simultaneously in split view. On compact width devices, the master view controller will be shown by default as the app launches.