Pull JSON data into detail VC and pass in existing data from main VC
Had to change the API call to one that only gives daily data as the API provider is giving an internal error for the real-time data. Passed in the price and ticker symbol via segue from the main VC. Set viewWillAppear to then request the JSON data function using the passed-in ticker symbol from the main VC tableView. Parsed the JSON data in the detail VC into existing models and set that data to display onscreen. Will work next on making the interface look functional for both iPad and iPhone.